Add new utils

+++ eslint.config.mjs
@@ -24,7 +24,7 @@ export default [
       '@typescript-eslint/no-unused-vars': 'warn',
       'no-console': 'warn',
       semi: ['error', 'always'],
-      quotes: ['error', 'double'],
+      quotes: ['error', 'single'],
       'prettier/prettier': 'error',
     },
   },
